move 'brain', 'config', 'runner' and 'template' to a package.
which makes them callable from external code
